requirements.txt
# API and Data Processing
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0

# Airflow
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.10.0

# dbt
dbt-core==1.7.4
dbt-postgres==1.7.4

# CrewAI and AI
crewai==0.1.26
langchain==0.1.0
langchain-community==0.0.10

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Testing
pytest==7.4.3
